requirements.txt
flask
requests
pytrends
flask_cors
functools
flasgger
flask_sqlalchemy
config
logging
marshmallow
python-dotenv
mysqlclient
SQLAlchemy
marshmallow-sqlalchemy
Flask-SQLAlchemy
smtplib
jinja2
pyotp